backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4
pyjwt>=2.10.1
python-multipart>=0.0.9
requests>=2.31.0
PyPDF2>=3.0.0
python-docx>=1.1.0
openai>=1.0.0
